#25 Fix logging for request error
