+++ main.go
@@ -4,17 +4,18 @@ import (
 	"github.com/labstack/echo/v4"
 	"github.com/ybkuroki/go-webapp-sample/config"
 	"github.com/ybkuroki/go-webapp-sample/logger"
+	"github.com/ybkuroki/go-webapp-sample/middleware"
 	"github.com/ybkuroki/go-webapp-sample/migration"
 	"github.com/ybkuroki/go-webapp-sample/repository"
 	"github.com/ybkuroki/go-webapp-sample/router"
-	"github.com/ybkuroki/go-webapp-sample/session"
 )
 
 func main() {
 	e := echo.New()
 
 	config.Load()
-	logger.InitLogger(e, config.GetConfig())
+	logger.InitLogger(config.GetConfig())
+	middleware.InitLoggerMiddleware(e)
 	e.Logger.Info("Loaded this configuration : application." + *config.GetEnv() + ".yml")
 
 	repository.InitDB()
@@ -24,7 +25,7 @@ func main() {
 	migration.InitMasterData(config.GetConfig())
 
 	router.Init(e, config.GetConfig())
-	session.Init(e, config.GetConfig())
+	middleware.InitSessionMiddleware(e, config.GetConfig())
 	if err := e.Start(":8080"); err != nil {
 		e.Logger.Error(err)
 	}
